"Fixed" a bug
Runtime errors made the program exit early, leaving the current
callframe dangling. On the next execution (in the REPL), a RET
instruction would resume execution of the dangling callframe, which is
an unintended behavior. This fix simply clears the stack and the
callstack after an error in the REPL. Such a fix is not necessary after
running a file since the programs exits completely after an error